6363 | 8 files | M DumpSource2/schemas/client/CDOTA_Modifier_EarthSpirit_GeomagneticGrip.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Furion_SpiritOfTheForest.h,  M DumpSource2/schemas/server/CDOTA_Modifier_EarthSpirit_GeomagneticGrip.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Furion_SpiritOfTheForest.h,  M game/bin/built_from_cl.txt,  M game/dota/bin/linuxsteamrt64/libclient_strings.txt,  M game/dota/pak01_dir.txt,  M game/dota/steam.inf
https://steamdb.info/patchnotes/17786216/

+++ DumpSource2/schemas/server/CDOTA_Modifier_EarthSpirit_GeomagneticGrip.h
@@ -1,7 +1,7 @@
 class CDOTA_Modifier_EarthSpirit_GeomagneticGrip
 {
-	int32 radius;
-	int32 rock_damage;
+	float32 radius;
+	float32 rock_damage;
 	float32 pull_units_per_second;
 	float32 pull_units_per_second_heroes;
 	float32 total_pull_distance;
