+++ DumpSource2/schemas/server/CDOTA_Modifier_Furion_SpiritOfTheForest.h
@@ -1,8 +1,8 @@
 class CDOTA_Modifier_Furion_SpiritOfTheForest
 {
 	float32 damage_per_tree_pct;
-	int32 radius;
+	float32 radius;
 	float32 multiplier;
-	int32 radius_treant;
+	float32 radius_treant;
 	int32 m_nTreeAmount;
 };
